requirements.txt
click==6.7
cloudpickle==0.5.3
ipdb==0.11
ipywidgets==7.3.0
jupyter==1.0.0
jupyterlab==0.32.1
lightgbm==2.1.1
matplotlib==2.2.2
notebook==5.5.0
numpy==1.15.2
pandas==0.22.0
PyYAML==3.13
reverse-geocoder==1.5.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.8.1
